+++ src/loginproviders/Linkedin.php
@@ -72,26 +72,21 @@ class Linkedin extends LoginProvider
 
     }
 
-    // Protected Methods
-    // =========================================================================
-
     /**
      * Returns the login provider’s OAuth provider.
      *
      * @return \League\OAuth2\Client\Provider\LinkedIn
+     * @throws \yii\base\InvalidConfigException
      */
-    protected function getOauthProvider()
+    public function getOauthProvider(): \League\OAuth2\Client\Provider\LinkedIn
     {
-        $providerInfos = $this->getInfos();
-
-        $config = [
-            'clientId' => (isset($providerInfos['clientId']) ? $providerInfos['clientId'] : ''),
-            'clientSecret' => (isset($providerInfos['clientSecret']) ? $providerInfos['clientSecret'] : ''),
-            'redirectUri' => $this->getRedirectUri(),
-        ];
+        $config = $this->getOauthProviderConfig();
 
         return new \League\OAuth2\Client\Provider\LinkedIn($config);
     }
+
+    // Protected Methods
+    // =========================================================================
 
     /**
      * @inheritdoc